# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-oriented Go Web API scaffold. Application error codes: 0 success; 10000-19999 common/input; 20000-29999 authentication/authorization; 30000-39999 business; 50000-59999 infrastructure.",
        "title": "Go API Template",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/route-policies/get": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "route-policies"
                ],
                "summary": "Get one discovered route and its authorization policy",
                "parameters": [
                    {
                        "description": "Route identifier",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.GetRoutePolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/routepolicy.Detail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/route-policies/page": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "route-policies"
                ],
                "summary": "Page discovered routes and their database-owned authorization policies",
                "parameters": [
                    {
                        "description": "Route policy filters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.PageRoutePoliciesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.RoutePolicyPage"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/route-policies/set": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "route-policies"
                ],
                "summary": "Create or update a route authorization policy",
                "parameters": [
                    {
                        "description": "Compiled CEL policy and permission references",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.SetRoutePolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/routepolicy.Detail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httptransport.GetRoutePolicyRequest": {
            "type": "object",
            "required": [
                "route_id"
            ],
            "properties": {
                "route_id": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.PageRoutePoliciesRequest": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "policy_status": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "route_status": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.Response": {
            "type": "object",
            "properties": {
                "body": {},
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.RoutePolicyPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routepolicy.Detail"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "httptransport.SetRoutePolicyRequest": {
            "type": "object",
            "required": [
                "expression",
                "route_id",
                "status"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "expected_version": {
                    "type": "integer"
                },
                "expression": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routepolicy.PermissionInput"
                    }
                },
                "route_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "routepolicy.Detail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "expression": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routepolicy.PermissionInput"
                    }
                },
                "policy_id": {
                    "type": "string"
                },
                "policy_status": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "route_id": {
                    "type": "string"
                },
                "route_status": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "routepolicy.PermissionInput": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Enter \"Bearer {JWT}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "PSK": {
            "description": "Enter \"PSK {shared-key}\" for routes configured with PSK authentication.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}